AMBARI-25290: Custom timeout ignored, python script has been killed due to timeout after waiting 2700 secs during upgrade (#3468)

+++ ambari-server/src/main/java/org/apache/ambari/server/controller/internal/ClusterStackVersionResourceProvider.java
@@ -687,7 +687,7 @@ public class ClusterStackVersionResourceProvider extends AbstractControllerResou
         INSTALL_PACKAGES_ACTION, Collections.singletonList(filter), roleParams);
 
     actionContext.setStackId(repoVersion.getStackId());
-    actionContext.setTimeout(Short.valueOf(configuration.getDefaultAgentTaskTimeout(true)));
+    actionContext.setTimeout(Integer.valueOf(configuration.getDefaultAgentTaskTimeout(true)));
 
     repoVersionHelper.addCommandRepositoryToContext(actionContext, repoVersion, osEntity);
